experimentation/requirements.txt
flask
openai
PyYAML
pytesseract
Pillow